# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Reusable Components",
        "description": "Reusable Components Api Docs",
        "version": "1"
    },
    "host": "reusable.applaurels.com",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/Delete_list": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Delete_list",
                "description": "delete chat thread",
                "operationId": "index_post",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Accesstoken",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "users",
                        "in": "formData",
                        "description": "users id as string eg <1,2,3,4>",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Parameter missing"
                    },
                    "200": {
                        "description": "Success"
                    },
                    "206": {
                        "description": "Unauthorized request"
                    },
                    "207": {
                        "description": "Header is missing"
                    }
                }
            }
        },
        "/Delete_message": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Delete_message",
                "description": "delete chat message",
                "operationId": "index_post",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Accesstoken",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "message_id",
                        "in": "formData",
                        "description": "message id to delete",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "sender_id",
                        "in": "formData",
                        "description": "sender id of that message ",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Parameter missing"
                    },
                    "200": {
                        "description": "Message deleted successfully"
                    },
                    "206": {
                        "description": "Unauthorized request"
                    },
                    "207": {
                        "description": "Header is missing"
                    }
                }
            }
        },
        "/Forgot": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Forgot Password",
                "description": "Forgot Password",
                "operationId": "index_post",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Forgot Email Send Success"
                    },
                    "206": {
                        "description": "Unauthorized request"
                    },
                    "207": {
                        "description": "Header is missing"
                    },
                    "211": {
                        "description": "Email Send failed"
                    },
                    "302": {
                        "description": "Email in not registered"
                    },
                    "418": {
                        "description": "Required Parameter Missing or Invalid"
                    }
                }
            }
        },
        "/Login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login Information",
                "description": "Login Information",
                "operationId": "index_post",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "device_id",
                        "in": "formData",
                        "description": "Unique Device Id",
                        "type": "string"
                    },
                    {
                        "name": "device_token",
                        "in": "formData",
                        "description": "Device Token required to send push",
                        "type": "string"
                    },
                    {
                        "name": "platform",
                        "in": "formData",
                        "description": "1: Android and 2: iOS",
                        "type": "string"
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Account Blocked"
                    },
                    "200": {
                        "description": "Login Success"
                    },
                    "206": {
                        "description": "Unauthorized request"
                    },
                    "207": {
                        "description": "Header is missing"
                    },
                    "418": {
                        "description": "Required Parameter Missing or Invalid"
                    }
                }
            }
        },
...
        "/Mark_read": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Mark as read",
                "description": "Mark message as read",
                "operationId": "index_post",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Accesstoken",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "users",
                        "in": "formData",
                        "description": "users eg<1,2,3>",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Parameter missing"
                    },
                    "200": {
                        "description": "Chat read successfully"
                    },
                    "206": {
                        "description": "Unauthorized request"
                    },
                    "207": {
                        "description": "Header is missing"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "basicAuth": {
            "type": "basic",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "basicAuth": [
                ""
            ]
        }
    ],
...
}